requirements.txt
einops==0.8.0
matplotlib==3.9.2
numpy==2.1.3
opencv-python==4.9.0.80
pandas==2.2.3
pillow==11.0.0
pydicom==3.0.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
tensorboard==2.18.0
timm==0.4.12
tokenizers==0.20.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
transformers==4.45.2
